These changes simplify `MergeQueue` and improve its efficiency.  Thank you to @mteldridge for noticing this issue.

Connects https://github.com/locationtech/geotrellis/issues/2358