@metasim 2017-10-02
It would be interesting to see how this method now performs.

https://github.com/locationtech/geotrellis/blob/e56e5e72463c5438e7c14b5baaba3527409d3160/bench/src/main/scala/geotrellis/spark/io/index/MergeQueueBench.scala#L58

Old runs are in `bench/archive`, but in CSV format.